feat: mv encoder to encode.go
