+++ prototype/prototype/encode.go
@@ -1,22 +1,12 @@
 // Copyright 2010 The Go Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
-
-// Package json implements encoding and decoding of JSON as defined in
-// RFC 7159. The mapping between JSON and Go values is described
-// in the documentation for the Marshal and Unmarshal functions.
-//
-// See "JSON and Go" for an introduction to this package:
-// https://golang.org/doc/articles/json_and_go.html
 package prototype
 
 import (
 	"encoding"
 	"fmt"
-	"github.com/go-leo/gox/convx"
 	"github.com/go-leo/gox/stringx"
-	"golang.org/x/exp/slices"
-	"math"
 	"reflect"
 	"sort"
 	"strconv"
@@ -25,135 +15,6 @@ import (
 	"unicode"
 )
 
-// Marshal returns the JSON encoding of v.
-//
-// Marshal traverses the value v recursively.
-// If an encountered value implements the Marshaler interface
-// and is not a nil pointer, Marshal calls its MarshalJSON method
-// to produce JSON. If no MarshalJSON method is present but the
-// value implements encoding.TextMarshaler instead, Marshal calls
-// its MarshalText method and encodes the result as a JSON string.
-// The nil pointer exception is not strictly necessary
-// but mimics a similar, necessary exception in the behavior of
-// UnmarshalJSON.
-//
-// Otherwise, Marshal uses the following type-dependent default encodings:
-//
-// Boolean values encode as JSON booleans.
-//
-// Floating point, integer, and Number values encode as JSON numbers.
-//
-// String values encode as JSON strings coerced to valid UTF-8,
-// replacing invalid bytes with the Unicode replacement rune.
-// So that the JSON will be safe to embed inside HTML <script> tags,
-// the string is encoded using HTMLEscape,
-// which replaces "<", ">", "&", U+2028, and U+2029 are escaped
-// to "\u003c","\u003e", "\u0026", "\u2028", and "\u2029".
-// This replacement can be disabled when using an Encoder,
-// by calling SetEscapeHTML(false).
-//
-// Array and slice values encode as JSON arrays, except that
-// []byte encodes as a base64-encoded string, and a nil slice
-// encodes as the null JSON value.
-//
-// Struct values encode as JSON objects.
-// Each exported struct field becomes a member of the object, using the
-// field name as the object key, unless the field is omitted for one of the
-// reasons given below.
-//
-// The encoding of each struct field can be customized by the format string
-// stored under the "json" key in the struct field's tag.
-// The format string gives the name of the field, possibly followed by a
-// comma-separated list of options. The name may be empty in order to
-// specify options without overriding the default field name.
-//
-// The "omitempty" option specifies that the field should be omitted
-// from the encoding if the field has an empty value, defined as
-// false, 0, a nil pointer, a nil interface value, and any empty array,
-// slice, map, or string.
-//
-// As a special case, if the field tag is "-", the field is always omitted.
-// Note that a field with name "-" can still be generated using the tag "-,".
-//
-// Examples of struct field tags and their meanings:
-//
-//	// Field appears in JSON as key "myName".
-//	Field int `json:"myName"`
-//
-//	// Field appears in JSON as key "myName" and
-//	// the field is omitted from the object if its value is empty,
-//	// as defined above.
-//	Field int `json:"myName,omitempty"`
-//
-//	// Field appears in JSON as key "Field" (the default), but
-//	// the field is skipped if empty.
-//	// Note the leading comma.
-//	Field int `json:",omitempty"`
-//
-//	// Field is ignored by this package.
-//	Field int `json:"-"`
-//
-//	// Field appears in JSON as key "-".
-//	Field int `json:"-,"`
-//
-// The "string" option signals that a field is stored as JSON inside a
-// JSON-encoded string. It applies only to fields of string, floating point,
-// integer, or boolean types. This extra level of encoding is sometimes used
-// when communicating with JavaScript programs:
-//
-//	Int64String int64 `json:",string"`
-//
-// The key name will be used if it's a non-empty string consisting of
-// only Unicode letters, digits, and ASCII punctuation except quotation
-// marks, backslash, and comma.
-//
-// Anonymous struct fields are usually marshaled as if their inner exported fields
-// were fields in the outer struct, subject to the usual Go visibility rules amended
-// as described in the next paragraph.
-// An anonymous struct field with a name given in its JSON tag is treated as
-// having that name, rather than being anonymous.
-// An anonymous struct field of interface type is treated the same as having
-// that type as its name, rather than being anonymous.
-//
-// The Go visibility rules for struct fields are amended for JSON when
-// deciding which field to marshal or unmarshal. If there are
-// multiple fields at the same level, and that level is the least
-// nested (and would therefore be the nesting level selected by the
-// usual Go rules), the following extra rules apply:
-//
-// 1) Of those fields, if any are JSON-tagged, only tagged fields are considered,
-// even if there are multiple untagged fields that would otherwise conflict.
-//
-// 2) If there is exactly one field (tagged or not according to the first rule), that is selected.
-//
-// 3) Otherwise there are multiple fields, and all are ignored; no error occurs.
-//
-// Handling of anonymous struct fields is new in Go 1.1.
-// Prior to Go 1.1, anonymous struct fields were ignored. To force ignoring of
-// an anonymous struct field in both current and earlier versions, give the field
-// a JSON tag of "-".
-//
-// Map values encode as JSON objects. The map's key type must either be a
-// string, an integer type, or implement encoding.TextMarshaler. The map keys
-// are sorted and used as JSON object keys by applying the following rules,
-// subject to the UTF-8 coercion described for string values above:
-//   - keys of any string type are used directly
-//   - encoding.TextMarshalers are marshaled
-//   - integer keys are converted to strings
-//
-// Pointer values encode as the value pointed to.
-// A nil pointer encodes as the null JSON value.
-//
-// Interface values encode as the value contained in the interface.
-// A nil interface value encodes as the null JSON value.
-//
-// Channel, complex, and function values cannot be encoded in JSON.
-// Attempting to encode such a value causes Marshal to return
-// an UnsupportedTypeError.
-//
-// JSON cannot represent cyclic data structures and Marshal does not
-// handle them. Passing cyclic structures to Marshal will result in
-// an error.
 func Marshal(v any) ([]byte, error) {
 	e := newEncodeState()
 	defer encodeStatePool.Put(e)
@@ -165,55 +26,6 @@ func Marshal(v any) ([]byte, error) {
 	return []byte{}, nil
 }
 
-// Marshaler is the interface implemented by types that
-// can marshal themselves into valid JSON.
-type Marshaler interface {
-	MarshalJSON() ([]byte, error)
-}
-
-// An UnsupportedTypeError is returned by Marshal when attempting
-// to encode an unsupported value type.
-type UnsupportedTypeError struct {
-	Type reflect.Type
-}
-
-func (e *UnsupportedTypeError) Error() string {
-	return "json: unsupported type: " + e.Type.String()
-}
-
-// An UnsupportedValueError is returned by Marshal when attempting
-// to encode an unsupported value.
-type UnsupportedValueError struct {
-	Value reflect.Value
-	Str   string
-}
-
-func (e *UnsupportedValueError) Error() string {
-	return "json: unsupported value: " + e.Str
-}
-
-// A MarshalerError represents an error from calling a MarshalJSON or MarshalText method.
-type MarshalerError struct {
-	Type       reflect.Type
-	Err        error
-	sourceFunc string
-}
-
-func (e *MarshalerError) Error() string {
-	srcFunc := e.sourceFunc
-	if srcFunc == "" {
-		srcFunc = "MarshalJSON"
-	}
-	return "json: error calling " + srcFunc +
-		" for type " + e.Type.String() +
-		": " + e.Err.Error()
-}
-
-// Unwrap returns the underlying error.
-func (e *MarshalerError) Unwrap() error { return e.Err }
-
-var hex = "0123456789abcdef"
-
 // An encodeState encodes JSON into a bytes.Buffer.
 type encodeState struct {
 	// Keep track of what pointers we've seen in the current recursive call
@@ -224,8 +36,6 @@ type encodeState struct {
 	ptrLevel uint
 	ptrSeen  map[any]struct{}
 }
-
-const startDetectingCyclesAfter = 1000
 
 var encodeStatePool sync.Pool
 
@@ -240,11 +50,6 @@ func newEncodeState() *encodeState {
 	}
 	return &encodeState{ptrSeen: make(map[any]struct{})}
 }
-
-// jsonError is an error wrapper type for internal use only.
-// Panics with errors are wrapped in jsonError so that the top-level recover
-// can distinguish intentional panics from this package.
-type jsonError struct{ error }
 
 func (e *encodeState) marshal(v any, opts encOpts) (err error) {
 	defer func() {
@@ -261,23 +66,9 @@ func (e *encodeState) marshal(v any, opts encOpts) (err error) {
 	return err
 }
 
-func isEmptyValue(v reflect.Value) bool {
-	switch v.Kind() {
-	case reflect.Array, reflect.Map, reflect.Slice, reflect.String:
-		return v.Len() == 0
-	case reflect.Bool:
-		return !v.Bool()
-	case reflect.Int, reflect.Int8, reflect.Int16, reflect.Int32, reflect.Int64:
-		return v.Int() == 0
-	case reflect.Uint, reflect.Uint8, reflect.Uint16, reflect.Uint32, reflect.Uint64, reflect.Uintptr:
-		return v.Uint() == 0
-	case reflect.Float32, reflect.Float64:
-		return v.Float() == 0
-	case reflect.Interface, reflect.Pointer:
-		return v.IsNil()
-	}
-	return false
-}
+var hex = "0123456789abcdef"
+
+const startDetectingCyclesAfter = 1000
 
 type encOpts struct {
 	// quoted causes primitive fields to be encoded inside JSON strings.
@@ -377,429 +168,6 @@ func newTypeEncoder(t reflect.Type, allowAddr bool) encoderFunc {
 	default:
 		return unsupportedTypeEncoder
 	}
-}
-
-func invalidValueEncoder(e *encodeState, fks []string, v reflect.Value, opts encOpts) error {
-	writeNull(e, fks, v, opts)
-	return nil
-}
-
-func writeNull(e *encodeState, fks []string, v reflect.Value, _ encOpts) {
-	printValue(fks, "null")
-}
-
-func printValue(fks []string, val any) {
-	fmt.Println(strings.Join(fks, "."), "=", val)
-}
-
-func marshalerEncoder(e *encodeState, fks []string, v reflect.Value, opts encOpts) error {
-	if v.Kind() == reflect.Pointer && v.IsNil() {
-		writeNull(e, fks, v, opts)
-		return nil
-	}
-	m, ok := v.Interface().(Marshaler)
-	if !ok {
-		writeNull(e, fks, v, opts)
-		return nil
-	}
-	b, err := m.MarshalJSON()
-	if err != nil {
-		return &MarshalerError{v.Type(), err, "MarshalJSON"}
-	}
-	printValue(fks, string(b))
-	return nil
-}
-
-func addrMarshalerEncoder(e *encodeState, fks []string, v reflect.Value, opts encOpts) error {
-	va := v.Addr()
-	if va.IsNil() {
-		writeNull(e, fks, v, opts)
-		return nil
-	}
-	m := va.Interface().(Marshaler)
-	b, err := m.MarshalJSON()
-	if err != nil {
-		return &MarshalerError{v.Type(), err, "MarshalJSON"}
-	}
-	printValue(fks, string(b))
-	return nil
-}
-
-func textMarshalerEncoder(e *encodeState, fks []string, v reflect.Value, opts encOpts) error {
-	if v.Kind() == reflect.Pointer && v.IsNil() {
-		writeNull(e, fks, v, opts)
-		return nil
-	}
-	m, ok := v.Interface().(encoding.TextMarshaler)
-	if !ok {
-		writeNull(e, fks, v, opts)
-		return nil
-	}
-	b, err := m.MarshalText()
-	if err != nil {
-		return &MarshalerError{v.Type(), err, "MarshalText"}
-	}
-	printValue(fks, string(b))
-	return nil
-}
-
-func addrTextMarshalerEncoder(e *encodeState, fks []string, v reflect.Value, opts encOpts) error {
-	va := v.Addr()
-	if va.IsNil() {
-		writeNull(e, fks, v, opts)
-		return nil
-	}
-	m := va.Interface().(encoding.TextMarshaler)
-	b, err := m.MarshalText()
-	if err != nil {
-		return &MarshalerError{v.Type(), err, "MarshalText"}
-	}
-	printValue(fks, string(b))
-	return nil
-}
-
-func boolEncoder(e *encodeState, fks []string, v reflect.Value, opts encOpts) error {
-	printValue(fks, v.Bool())
-	return nil
-}
-
-func intEncoder(e *encodeState, fks []string, v reflect.Value, opts encOpts) error {
-	printValue(fks, v.Int())
-	return nil
-}
-
-func uintEncoder(e *encodeState, fks []string, v reflect.Value, opts encOpts) error {
-	printValue(fks, v.Uint())
-	return nil
-}
-
-type floatEncoder int // number of bits
-
-func (bits floatEncoder) encode(e *encodeState, fks []string, v reflect.Value, opts encOpts) error {
-	f := v.Float()
-	if math.IsInf(f, 0) || math.IsNaN(f) {
-		return &UnsupportedValueError{v, strconv.FormatFloat(f, 'g', -1, int(bits))}
-	}
-	printValue(fks, v.Float())
-	return nil
-}
-
-var (
-	float32Encoder = (floatEncoder(32)).encode
-	float64Encoder = (floatEncoder(64)).encode
-)
-
-func stringEncoder(e *encodeState, fks []string, v reflect.Value, opts encOpts) error {
-	if v.Type() == numberType {
-		numStr := v.String()
-		// In Go1.5 the empty string encodes to "0", while this is not a valid number literal
-		// we keep compatibility so check validity after this.
-		if numStr == "" {
-			numStr = "0" // Number's zero-val
-		}
-		if !isValidNumber(numStr) {
-			return fmt.Errorf("json: invalid number literal %q", numStr)
-		}
-		printValue(fks, v.String())
-		return nil
-	}
-	if opts.quoted {
-		printValue(fks, v.String())
-		return nil
-	}
-	printValue(fks, v.String())
-	return nil
-}
-
-// isValidNumber reports whether s is a valid JSON number literal.
-func isValidNumber(s string) bool {
-	// This function implements the JSON numbers grammar.
-	// See https://tools.ietf.org/html/rfc7159#section-6
-	// and https://www.json.org/img/number.png
-
-	if s == "" {
-		return false
-	}
-
-	// Optional -
-	if s[0] == '-' {
-		s = s[1:]
-		if s == "" {
-			return false
-		}
-	}
-
-	// Digits
-	switch {
-	default:
-		return false
-
-	case s[0] == '0':
-		s = s[1:]
-
-	case '1' <= s[0] && s[0] <= '9':
-		s = s[1:]
-		for len(s) > 0 && '0' <= s[0] && s[0] <= '9' {
-			s = s[1:]
-		}
-	}
-
-	// . followed by 1 or more digits.
-	if len(s) >= 2 && s[0] == '.' && '0' <= s[1] && s[1] <= '9' {
-		s = s[2:]
-		for len(s) > 0 && '0' <= s[0] && s[0] <= '9' {
-			s = s[1:]
-		}
-	}
-
-	// e or E followed by an optional - or + and
-	// 1 or more digits.
-	if len(s) >= 2 && (s[0] == 'e' || s[0] == 'E') {
-		s = s[1:]
-		if s[0] == '+' || s[0] == '-' {
-			s = s[1:]
-			if s == "" {
-				return false
-			}
-		}
-		for len(s) > 0 && '0' <= s[0] && s[0] <= '9' {
-			s = s[1:]
-		}
-	}
-
-	// Make sure we are at the end.
-	return s == ""
-}
-
-func interfaceEncoder(e *encodeState, fks []string, v reflect.Value, opts encOpts) error {
-	if v.IsNil() {
-		writeNull(e, fks, v, opts)
-		return nil
-	}
-	value := v.Elem()
-	return valueEncoder(value)(e, fks, value, opts)
-}
-
-func unsupportedTypeEncoder(e *encodeState, fks []string, v reflect.Value, _ encOpts) error {
-	return &UnsupportedTypeError{v.Type()}
-}
-
-type structEncoder struct {
-	fields structFields
-}
-
-type structFields struct {
-	list      []field
-	nameIndex map[string]int
-}
-
-func (se structEncoder) encode(e *encodeState, fks []string, v reflect.Value, opts encOpts) error {
-FieldLoop:
-	for i := range se.fields.list {
-		f := &se.fields.list[i]
-
-		// Find the nested struct field by following f.index.
-		fv := v
-		for _, i := range f.index {
-			if fv.Kind() == reflect.Pointer {
-				if fv.IsNil() {
-					continue FieldLoop
-				}
-				fv = fv.Elem()
-			}
-			fv = fv.Field(i)
-		}
-
-		if f.omitEmpty && isEmptyValue(fv) {
-			continue
-		}
-		opts.quoted = f.quoted
-		if err := f.encoder(e, append(fks, f.name), fv, opts); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-func newStructEncoder(t reflect.Type) encoderFunc {
-	se := structEncoder{fields: cachedTypeFields(t)}
-	return se.encode
-}
-
-type mapEncoder struct {
-	elemEnc encoderFunc
-}
-
-func (me mapEncoder) encode(e *encodeState, fks []string, v reflect.Value, opts encOpts) error {
-	if v.IsNil() {
-		writeNull(e, fks, v, opts)
-		return nil
-	}
-	if e.ptrLevel++; e.ptrLevel > startDetectingCyclesAfter {
-		// We're a large number of nested ptrEncoder.encode calls deep;
-		// start checking if we've run into a pointer cycle.
-		ptr := v.UnsafePointer()
-		if _, ok := e.ptrSeen[ptr]; ok {
-			return &UnsupportedValueError{v, fmt.Sprintf("encountered a cycle via %s", v.Type())}
-		}
-		e.ptrSeen[ptr] = struct{}{}
-		defer delete(e.ptrSeen, ptr)
-	}
-
-	// Extract and sort the keys.
-	sv := make([]reflectWithString, v.Len())
-	mi := v.MapRange()
-	for i := 0; mi.Next(); i++ {
-		sv[i].k = mi.Key()
-		sv[i].v = mi.Value()
-		if err := sv[i].resolve(); err != nil {
-			return fmt.Errorf("json: encoding error for type %q: %q", v.Type().String(), err.Error())
-		}
-	}
-	sort.Slice(sv, func(i, j int) bool { return sv[i].ks < sv[j].ks })
-
-	for _, kv := range sv {
-		if err := me.elemEnc(e, append(slices.Clone(fks), kv.ks), kv.v, opts); err != nil {
-			return err
-		}
-	}
-	e.ptrLevel--
-	return nil
-}
-
-func newMapEncoder(t reflect.Type) encoderFunc {
-	switch t.Key().Kind() {
-	case reflect.String,
-		reflect.Int, reflect.Int8, reflect.Int16, reflect.Int32, reflect.Int64,
-		reflect.Uint, reflect.Uint8, reflect.Uint16, reflect.Uint32, reflect.Uint64, reflect.Uintptr:
-	default:
-		if !t.Key().Implements(textMarshalerType) {
-			return unsupportedTypeEncoder
-		}
-	}
-	me := mapEncoder{typeEncoder(t.Elem())}
-	return me.encode
-}
-
-func encodeByteSlice(e *encodeState, fks []string, v reflect.Value, opts encOpts) error {
-	if v.IsNil() {
-		writeNull(e, fks, v, opts)
-		return nil
-	}
-	printValue(fks, string(v.Bytes()))
-	return nil
-}
-
-// sliceEncoder just wraps an arrayEncoder, checking to make sure the value isn't nil.
-type sliceEncoder struct {
-	arrayEnc encoderFunc
-}
-
-func (se sliceEncoder) encode(e *encodeState, fks []string, v reflect.Value, opts encOpts) error {
-	if v.IsNil() {
-		writeNull(e, fks, v, opts)
-		return nil
-	}
-	if e.ptrLevel++; e.ptrLevel > startDetectingCyclesAfter {
-		// We're a large number of nested ptrEncoder.encode calls deep;
-		// start checking if we've run into a pointer cycle.
-		// Here we use a struct to memorize the pointer to the first element of the slice
-		// and its length.
-		ptr := struct {
-			ptr interface{} // always an unsafe.Pointer, but avoids a dependency on package unsafe
-			len int
-		}{v.UnsafePointer(), v.Len()}
-		if _, ok := e.ptrSeen[ptr]; ok {
-			return &UnsupportedValueError{v, fmt.Sprintf("encountered a cycle via %s", v.Type())}
-		}
-		e.ptrSeen[ptr] = struct{}{}
-		defer delete(e.ptrSeen, ptr)
-	}
-	if err := se.arrayEnc(e, fks, v, opts); err != nil {
-		return err
-	}
-	e.ptrLevel--
-	return nil
-}
-
-func newSliceEncoder(t reflect.Type) encoderFunc {
-	// Byte slices get special treatment; arrays don't.
-	if t.Elem().Kind() == reflect.Uint8 {
-		p := reflect.PointerTo(t.Elem())
-		if !p.Implements(marshalerType) && !p.Implements(textMarshalerType) {
-			return encodeByteSlice
-		}
-	}
-	enc := sliceEncoder{newArrayEncoder(t)}
-	return enc.encode
-}
-
-type arrayEncoder struct {
-	elemEnc encoderFunc
-}
-
-func (ae arrayEncoder) encode(e *encodeState, fks []string, v reflect.Value, opts encOpts) error {
-	n := v.Len()
-	for i := 0; i < n; i++ {
-		if err := ae.elemEnc(e, append(slices.Clone(fks), convx.ToString(i)), v.Index(i), opts); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-func newArrayEncoder(t reflect.Type) encoderFunc {
-	enc := arrayEncoder{typeEncoder(t.Elem())}
-	return enc.encode
-}
-
-type ptrEncoder struct {
-	elemEnc encoderFunc
-}
-
-func (pe ptrEncoder) encode(e *encodeState, fks []string, v reflect.Value, opts encOpts) error {
-	if v.IsNil() {
-		writeNull(e, fks, v, opts)
-		return nil
-	}
-	if e.ptrLevel++; e.ptrLevel > startDetectingCyclesAfter {
-		// We're a large number of nested ptrEncoder.encode calls deep;
-		// start checking if we've run into a pointer cycle.
-		ptr := v.Interface()
-		if _, ok := e.ptrSeen[ptr]; ok {
-			return &UnsupportedValueError{v, fmt.Sprintf("encountered a cycle via %s", v.Type())}
-		}
-		e.ptrSeen[ptr] = struct{}{}
-		defer delete(e.ptrSeen, ptr)
-	}
-	if err := pe.elemEnc(e, fks, v.Elem(), opts); err != nil {
-		return err
-	}
-	e.ptrLevel--
-	return nil
-}
-
-func newPtrEncoder(t reflect.Type) encoderFunc {
-	enc := ptrEncoder{typeEncoder(t.Elem())}
-	return enc.encode
-}
-
-type condAddrEncoder struct {
-	canAddrEnc, elseEnc encoderFunc
-}
-
-func (ce condAddrEncoder) encode(e *encodeState, fks []string, v reflect.Value, opts encOpts) error {
-	if v.CanAddr() {
-		return ce.canAddrEnc(e, fks, v, opts)
-	}
-	return ce.elseEnc(e, fks, v, opts)
-}
-
-// newCondAddrEncoder returns an encoder that checks whether its value
-// CanAddr and delegates to canAddrEnc if so, else to elseEnc.
-func newCondAddrEncoder(canAddrEnc, elseEnc encoderFunc) encoderFunc {
-	enc := condAddrEncoder{canAddrEnc: canAddrEnc, elseEnc: elseEnc}
-	return enc.encode
 }
 
 func isValidTag(s string) bool {
