+++ prototype/prototype/marshalable.go
@@ -1,0 +1,7 @@
+package prototype
+
+// Marshaler is the interface implemented by types that
+// can marshal themselves into valid JSON.
+type Marshaler interface {
+	MarshalJSON() ([]byte, error)
+}
